Fixes modal sliders

+++ app/assets/javascripts/modal_slider.js
@@ -12,9 +12,9 @@ $( document ).ready(function() {
 
 	$( ".price-target" ).slider({
 	  animate: "fast",
-	  value: parseInt(price.dataset.price),
+	  value: parseFloat(price.dataset.price),
 	  min: 0,
-	  max: parseInt(price.dataset.price) + 2000,
+	  max: parseFloat(price.dataset.price) + 2000.00,
 	  range: true,
 	  step: .01
 	});
